refactor package script to allow for general linux packaging
and create target
